+++ src/pages/auth/SignUpPage.test.tsx
