+++ vite.config.ts
@@ -35,6 +35,7 @@ export default defineConfig({
         'src/vite-env.d.ts',
         'src/main.tsx',
         'src/App.tsx',
+        'src/**/*.test.{ts,tsx}',
       ],
     },
   },
