+++ vitest-setup.ts
@@ -1,6 +1,6 @@
+import { server } from '@/tests/fixtures/server/server';
 import '@testing-library/jest-dom/vitest';
 import { beforeAll, afterAll, afterEach } from 'vitest';
-import { server } from './src/tests/mocks/server';
 
 beforeAll(() => {
   server.listen();
